refactor(admin): extract ActivityLog, AnnouncementsAdmin, FeaturedCommunities
Also delete the dead blockInstance/unblockInstance stubs and their mock
state. AdminPanel.vue is down to 2.7k lines.
