+++ src/views/AdminPanel.vue
@@ -128,197 +128,15 @@
       <ReportsModeration />
 
       <!-- Recent Activity -->
-      <div class="admin-module activity-module">
-        <div class="module-header">
-          <Icon name="activity" :size="20" />
-          <h2>Recent Activity</h2>
-          <select v-model="activityFilter" class="cyber-select">
-            <option value="all">All Events</option>
-            <option value="federation">Federation</option>
-            <option value="moderation">Moderation</option>
-            <option value="security">Security</option>
-          </select>
-        </div>
-        <div class="activity-feed">
-          <div v-for="event in filteredRecentActivity" :key="event.id" class="activity-item">
-            <div class="activity-icon" :class="getActivityCategory(event.type)">
-              <Icon :name="getActivityIcon(event.type)" :size="16" />
-            </div>
-            <div class="activity-content">
-              <div class="activity-message">{{ formatActivityMessage(event) }}</div>
-              <div class="activity-meta">
-                <span class="activity-time">{{ formatTime(event.timestamp) }}</span>
-                <span class="activity-source">{{ event.source }}</span>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
+      <ActivityLog />
 
       <!-- Announcements -->
-      <div class="admin-module announcements-module">
-        <div class="module-header">
-          <Icon name="message-square" :size="20" />
-          <h2>Announcements</h2>
-          <button @click="showAnnouncementForm = true" class="primary-btn-sm">
-            <Icon name="plus" :size="14" />
-            Create
-          </button>
-        </div>
-        <p class="module-hint">Instance-wide announcements shown to users. Create and manage them here.</p>
-        <div v-if="showAnnouncementForm" class="announcement-form">
-          <h4>{{ editingAnnouncementId ? 'Edit' : 'New' }} Announcement</h4>
-          <div class="form-row">
-            <label>Title</label>
-            <input v-model="announcementForm.title" class="cyber-input" placeholder="Announcement title" />
-          </div>
-          <div class="form-row">
-            <label>Content (supports basic HTML)</label>
-            <textarea v-model="announcementForm.content" class="cyber-input" rows="4" placeholder="Announcement content"></textarea>
-          </div>
-          <div class="form-row">
-            <label>Icon (emoji or name: info, warning, megaphone)</label>
-            <input v-model="announcementForm.icon" class="cyber-input" placeholder="info" />
-          </div>
-          <div class="form-row">
-            <label>Image URL (optional)</label>
-            <input v-model="announcementForm.image_url" class="cyber-input" type="url" placeholder="https://..." />
-          </div>
-          <div class="form-row two-col">
-            <div>
-              <label>Starts at (optional)</label>
-              <input
-                v-model="announcementForm.starts_at"
-                class="cyber-input"
-                type="datetime-local"
-              />
-              <p class="form-hint">Leave empty to publish immediately. Times are interpreted in your local timezone.</p>
-            </div>
-            <div>
-              <label>Ends at (optional)</label>
-              <input
-                v-model="announcementForm.ends_at"
-                class="cyber-input"
-                type="datetime-local"
-                :min="announcementForm.starts_at || undefined"
-              />
-              <p class="form-hint">Leave empty for no expiration. After this time the announcement is hidden from users automatically.</p>
-            </div>
-          </div>
-          <div class="form-row checks">
-            <label class="toggle-label">
-              <input type="checkbox" v-model="announcementForm.is_pinned" />
-              <span class="toggle-slider"></span>
-              <span class="toggle-text">Pinned</span>
-            </label>
-            <label class="toggle-label">
-              <input type="checkbox" v-model="announcementForm.show_popup" />
-              <span class="toggle-slider"></span>
-              <span class="toggle-text">Show popup</span>
-            </label>
-            <label class="toggle-label" v-if="editingAnnouncementId">
-              <input type="checkbox" v-model="announcementForm.is_active" />
-              <span class="toggle-slider"></span>
-              <span class="toggle-text">Active</span>
-            </label>
-          </div>
-          <div class="form-actions">
-            <button @click="saveAnnouncement" class="primary-btn-sm" :disabled="!announcementForm.title || !announcementForm.content">
-              {{ editingAnnouncementId ? 'Update' : 'Create' }}
-            </button>
-            <button @click="cancelAnnouncementForm" class="cyber-btn-sm">Cancel</button>
-          </div>
-        </div>
-        <div class="announcements-list">
-          <div v-for="a in announcements" :key="a.id" class="announcement-item">
-            <div class="announcement-meta">
-              <span class="announcement-icon">{{ getAnnouncementIcon(a.icon) }}</span>
-              <span class="announcement-title">{{ a.title }}</span>
-              <span v-if="a.is_pinned" class="badge">Pinned</span>
-              <span v-if="!a.is_active" class="badge inactive">Inactive</span>
-            </div>
-            <div class="announcement-actions">
-              <button @click="editAnnouncement(a)" class="action-btn-sm" title="Edit">
-                <Icon name="edit" :size="14" />
-              </button>
-              <button @click="deleteAnnouncement(a.id)" class="danger-btn-sm" title="Delete">
-                <Icon name="trash" :size="14" />
-              </button>
-            </div>
-          </div>
-          <div v-if="announcements.length === 0 && !loadingStates.announcements" class="empty-state">
-            No announcements. Create one to notify users.
-          </div>
-        </div>
-      </div>
+      <AnnouncementsAdmin />
 
       <!-- Featured Communities -->
-      <div class="admin-module featured-module">
-        <div class="module-header">
-          <Icon name="star" :size="20" />
-          <h2>Featured Communities</h2>
-          <button @click="loadFeaturedServers" class="action-btn" :disabled="loadingStates.featuredServers">
-            <Icon v-if="loadingStates.featuredServers" name="loader" :size="16" class="spin" />
-            <Icon v-else name="refresh-cw" :size="16" />
-            Refresh
-          </button>
-        </div>
-        <p class="module-hint">Pin public servers to the top of the community discovery page. Featured servers appear first.</p>
-        <div v-if="loadingStates.featuredServers" class="loading-state">
-          <LoadingSpinner :size="20" />
-          <span>Loading communities...</span>
-        </div>
-        <div v-else class="featured-servers-list">
-          <div
-            v-for="server in featuredServersList"
-            :key="server.id"
-            class="featured-server-item"
-            :class="{ featured: server.is_featured }"
-          >
-            <div class="server-icon-wrap">
-              <img
-                :src="getServerIconUrl(server.icon)"
-                :alt="server.name"
-                class="server-icon"
-              />
-              <span v-if="server.is_featured" class="featured-badge">
-                <Icon name="star" :size="12" />
-              </span>
-            </div>
-            <div class="server-details">
-              <div class="server-name">{{ server.name }}</div>
-              <div class="server-meta">
-                {{ server.member_count ?? 0 }} members
-                <span v-if="server.is_featured" class="featured-order">#{{ server.featured_order }}</span>
-              </div>
-            </div>
-            <button
-              @click="toggleFeaturedServer(server)"
-              class="action-btn-sm"
-              :class="server.is_featured ? 'unpin-btn' : 'pin-btn'"
-              :disabled="featuredServerToggling.has(server.id)"
-              :title="server.is_featured ? 'Remove from featured' : 'Add to featured'"
-            >
-              <Icon v-if="featuredServerToggling.has(server.id)" name="loader" :size="14" class="spin" />
-              <Icon v-else :name="server.is_featured ? 'x' : 'star'" :size="14" />
-              {{ server.is_featured ? 'Unpin' : 'Pin' }}
-            </button>
-          </div>
-          <div v-if="featuredServersList.length === 0 && !loadingStates.featuredServers" class="empty-state">
-            No public communities yet. Create public servers to feature them.
-          </div>
-        </div>
-      </div>
+      <FeaturedCommunities />
 
-
-
-      <!-- Configuration -->
-      <InstanceConfig />
-
-      <!-- Funding Management -->
-      <FundingSupporters />
-
-      <!-- Performance Monitoring -->
+      <!-- Recent Activity -->
       <div class="admin-module performance-module">
         <div class="module-header">
           <Icon name="activity" :size="20" />
@@ -348,24 +166,19 @@ import { ref, computed, onMounted } from 'vue'
 import { debug } from '@/utils/debug'
 import { useAuthStore } from '@/stores/auth'
 import { useRouter } from 'vue-router'
-import { useToast } from 'vue-toastification'
 import Icon from '@/components/common/Icon.vue'
-import LoadingSpinner from '@/components/common/LoadingSpinner.vue'
 import EmojiImporter from '@/components/admin/EmojiImporter.vue'
 import PerformanceMonitoring from '@/components/admin/PerformanceMonitoring.vue'
 import FederationManagement from '@/components/admin/FederationManagement.vue'
 import UserManagement from '@/components/admin/UserManagement.vue'
 import ReportsModeration from '@/components/admin/ReportsModeration.vue'
-import InstanceConfig from '@/components/admin/InstanceConfig.vue'
-import FundingSupporters from '@/components/admin/FundingSupporters.vue'
-import { adminService, type AdminActivity } from '@/services/AdminService'
-import { announcementService, type Announcement } from '@/services/AnnouncementService'
-import { usePublicServersStore } from '@/stores/usePublicServers'
-import { getServerIconUrl } from '@/utils/serverUtils'
+import ActivityLog from '@/components/admin/ActivityLog.vue'
+import AnnouncementsAdmin from '@/components/admin/AnnouncementsAdmin.vue'
+import FeaturedCommunities from '@/components/admin/FeaturedCommunities.vue'
+import { adminService } from '@/services/AdminService'
 
 const authStore = useAuthStore()
 const router = useRouter()
-const toast = useToast()
 
 // Security check - only allow admins
 onMounted(async () => {
@@ -383,83 +196,14 @@ onMounted(async () => {
   }
 
   await loadInitialData()
-  window.addEventListener('admin:activity-changed', onActivityChanged)
 })
-
-const onActivityChanged = () => { void loadRecentActivity() }
 
 // Reactive data
 const loading = ref(false)
-const activityFilter = ref('all')
-const newBlockDomain = ref('')
-const newBlockReason = ref('')
 
 
 
-// Loading states
-const loadingStates = ref({
-  announcements: false,
-  featuredServers: false,
-})
 
-
-// Announcements
-const announcements = ref<Announcement[]>([])
-const showAnnouncementForm = ref(false)
-const editingAnnouncementId = ref<string | null>(null)
-const announcementForm = ref({
-  title: '',
-  content: '',
-  icon: 'info',
-  image_url: '',
-  is_pinned: false,
-  show_popup: true,
-  is_active: true,
-  // datetime-local strings ("YYYY-MM-DDTHH:mm" in the admin's local TZ).
-  // Empty string means "use the DB default" (now() for starts_at, NULL /
-  // never-expires for ends_at). Converted to ISO at save time via
-  // `localInputToIso` so PostgreSQL stores UTC.
-  starts_at: '',
-  ends_at: '',
-})
-
-// ---------------------------------------------------------------------------
-// datetime-local <-> ISO helpers
-// ---------------------------------------------------------------------------
-// HTML's `datetime-local` input gives us a naive local timestamp without a
-// timezone (e.g. "2026-05-26T21:00"). We convert in both directions so the
-// admin sees their own timezone but the DB stores UTC ISO strings.
-//
-// Both helpers are intentionally null-safe and treat an empty input as "no
-// value" rather than as the Unix epoch.
-function localInputToIso(local: string): string | undefined {
-  if (!local) return undefined
-  const d = new Date(local)
-  if (Number.isNaN(d.getTime())) return undefined
-  return d.toISOString()
-}
-
-function isoToLocalInput(iso: string | null | undefined): string {
-  if (!iso) return ''
-  const d = new Date(iso)
-  if (Number.isNaN(d.getTime())) return ''
-  // Shift by the local TZ offset so `toISOString().slice(0, 16)` formats
-  // as the admin's wall-clock time rather than as UTC.
-  const tzOffsetMs = d.getTimezoneOffset() * 60_000
-  return new Date(d.getTime() - tzOffsetMs).toISOString().slice(0, 16)
-}
-
-// Featured communities
-const featuredServersList = ref<Array<{
-  id: string;
-  name: string;
-  description?: string;
-  icon?: string;
-  is_featured: boolean;
-  featured_order: number;
-  member_count?: number;
-}>>([])
-const featuredServerToggling = ref<Set<string>>(new Set())
 
 // User servers modal
 
@@ -485,13 +229,6 @@ const systemHealth = ref({
   memory: { used: 0, total: '16GB' }
 })
 
-// Users data
-const blockedInstances = ref([
-  { domain: 'bad-instance.com', reason: 'Spam and harassment' },
-  { domain: 'another-bad.net', reason: 'Policy violations' }
-])
-
-const recentActivity = ref<any[]>([])
 
 // Computed properties
 const systemStatus = computed(() => {
@@ -524,15 +261,6 @@ const healthStatus = computed(() => {
   return { class: 'error', text: 'Critical Issues' }
 })
 
-// Filter recent activity by category and format JSON details for display
-const filteredRecentActivity = computed(() => {
-  let list = recentActivity.value
-  if (activityFilter.value !== 'all') {
-    list = list.filter(e => getActivityCategory(e.type) === activityFilter.value)
-  }
-  return list
-})
-
 
 
 // Methods
@@ -549,190 +277,11 @@ const loadInitialData = async () => {
     ])
 
     void Promise.allSettled([
-      loadRecentActivity(),
-      loadAnnouncements(),
-      loadFeaturedServers(),
     ])
   } catch (error) {
     debug.error('Failed to load admin data:', error)
   } finally {
     loading.value = false
-  }
-}
-
-const loadAnnouncements = async () => {
-  loadingStates.value.announcements = true
-  try {
-    announcements.value = await announcementService.getAllAnnouncements()
-  } catch (error) {
-    debug.error('Failed to load announcements:', error)
-    announcements.value = []
-  } finally {
-    loadingStates.value.announcements = false
-  }
-}
-
-const loadFeaturedServers = async () => {
-  loadingStates.value.featuredServers = true
-  try {
-    featuredServersList.value = await adminService.getPublicServersForAdmin()
-  } catch (error) {
-    debug.error('Failed to load featured servers:', error)
-    featuredServersList.value = []
-  } finally {
-    loadingStates.value.featuredServers = false
-  }
-}
-
-const toggleFeaturedServer = async (server: { id: string; name: string; is_featured: boolean; featured_order: number }) => {
-  if (featuredServerToggling.value.has(server.id)) return
-  featuredServerToggling.value = new Set([...featuredServerToggling.value, server.id])
-  try {
-    const newFeatured = !server.is_featured
-    const order = newFeatured
-      ? Math.max(0, ...featuredServersList.value.filter(s => s.is_featured).map(s => s.featured_order), -1) + 1
-      : 0
-    await adminService.setServerFeatured(server.id, newFeatured, order)
-    toast.success(newFeatured ? `Featured "${server.name}"` : `Removed "${server.name}" from featured`)
-    await loadFeaturedServers()
-    usePublicServersStore().fetchPublicServers(true)
-  } catch (error: any) {
-    debug.error('Failed to toggle featured:', error)
-    toast.error(error.message || 'Failed to update featured status')
-  } finally {
-    featuredServerToggling.value = new Set([...featuredServerToggling.value].filter(id => id !== server.id))
-  }
-}
-
-const getAnnouncementIcon = (icon: string | undefined) => {
-  if (!icon) return '📢'
-  const icons: Record<string, string> = {
-    info: 'ℹ️',
-    warning: '⚠️',
-    megaphone: '📢',
-  }
-  return icons[icon] || (icon.length <= 2 ? icon : '📢')
-}
-
-const saveAnnouncement = async () => {
-  if (!announcementForm.value.title || !announcementForm.value.content) return
-
-  // Asymmetric handling for the two timestamps:
-  //   * `starts_at`: an empty input means "use the DB default (now()) on
-  //     create" / "leave the existing value alone on update". Sending
-  //     undefined makes the Supabase client omit the key entirely.
-  //   * `ends_at`:   an empty input means "no expiration" - admins clearing
-  //     the field must be able to explicitly drop the value back to NULL,
-  //     otherwise an existing expiry would be impossible to remove. We
-  //     send `null` rather than undefined so the column is overwritten.
-  // Cross-field validation: end-before-start would silently produce an
-  // announcement that's already expired the moment it's published.
-  const startsIso = localInputToIso(announcementForm.value.starts_at)
-  const endsIsoOrNull = announcementForm.value.ends_at
-    ? localInputToIso(announcementForm.value.ends_at) ?? null
-    : null
-  if (startsIso && endsIsoOrNull && new Date(endsIsoOrNull) <= new Date(startsIso)) {
-    toast.error('"Ends at" must be after "Starts at"')
-    return
-  }
-  try {
-    if (editingAnnouncementId.value) {
-      await announcementService.updateAnnouncement(editingAnnouncementId.value, {
-        title: announcementForm.value.title,
-        content: announcementForm.value.content,
-        icon: announcementForm.value.icon || 'info',
-        image_url: announcementForm.value.image_url || undefined,
-        is_pinned: announcementForm.value.is_pinned,
-        show_popup: announcementForm.value.show_popup,
-        is_active: announcementForm.value.is_active,
-        starts_at: startsIso,
-        ends_at: endsIsoOrNull,
-      })
-      toast.success('Announcement updated')
-    } else {
-      await announcementService.createAnnouncement({
-        title: announcementForm.value.title,
-        content: announcementForm.value.content,
-        icon: announcementForm.value.icon || 'info',
-        image_url: announcementForm.value.image_url || undefined,
-        is_pinned: announcementForm.value.is_pinned,
-        show_popup: announcementForm.value.show_popup,
-        starts_at: startsIso,
-        ends_at: endsIsoOrNull,
-      })
-      toast.success('Announcement created')
-    }
-    cancelAnnouncementForm()
-    await loadAnnouncements()
-  } catch (error: any) {
-    debug.error('Failed to save announcement:', error)
-    toast.error(error.message || 'Failed to save announcement')
-  }
-}
-
-const cancelAnnouncementForm = () => {
-  showAnnouncementForm.value = false
-  editingAnnouncementId.value = null
-  announcementForm.value = {
-    title: '',
-    content: '',
-    icon: 'info',
-    image_url: '',
-    is_pinned: false,
-    show_popup: true,
-    is_active: true,
-    starts_at: '',
-    ends_at: '',
-  }
-}
-
-const editAnnouncement = (a: Announcement) => {
-  editingAnnouncementId.value = a.id
-  announcementForm.value = {
-    title: a.title,
-    content: a.content,
-    icon: a.icon || 'info',
-    image_url: a.image_url || '',
-    is_pinned: a.is_pinned ?? false,
-    show_popup: a.show_popup ?? true,
-    is_active: (a as any).is_active ?? true,
-    // Convert the DB-stored UTC ISO back to the admin's local "YYYY-MM-DDTHH:mm"
-    // so the datetime-local inputs render the wall-clock time they originally
-    // entered, not UTC. Empty when unset.
-    starts_at: isoToLocalInput((a as any).starts_at),
-    ends_at: isoToLocalInput(a.ends_at),
-  }
-  showAnnouncementForm.value = true
-}
-
-const deleteAnnouncement = async (id: string) => {
-  if (!confirm('Delete this announcement?')) return
-  try {
-    await announcementService.deleteAnnouncement(id)
-    toast.success('Announcement deleted')
-    await loadAnnouncements()
-  } catch (error: any) {
-    debug.error('Failed to delete announcement:', error)
-    toast.error(error.message || 'Failed to delete')
-  }
-}
-
-const loadRecentActivity = async () => {
-  try {
-    const activity = await adminService.getRecentActivity(20)
-    
-    recentActivity.value = activity.map((event: AdminActivity) => ({
-      id: event.id,
-      type: event.action_type,
-      message: event.details,
-      targetUsername: event.target_username,
-      timestamp: new Date(event.created_at),
-      source: `Admin: ${event.admin_username}`,
-      admin_id: event.admin_id
-    }))
-  } catch (error) {
-    debug.error('Failed to load recent activity:', error)
-    recentActivity.value = []
   }
 }
 
@@ -792,57 +341,6 @@ const exportLogs = () => {
   debug.log('Exporting logs...')
 }
 
-// eslint-disable-next-line unused-imports/no-unused-vars
-const blockInstance = async () => {
-  if (newBlockDomain.value && newBlockReason.value) {
-    try {
-      await adminService.moderateInstance(
-        newBlockDomain.value,
-        'block',
-        newBlockReason.value,
-        authStore.session?.user?.id || ''
-      )
-
-      blockedInstances.value.push({
-        domain: newBlockDomain.value,
-        reason: newBlockReason.value
-      })
-      
-      newBlockDomain.value = ''
-      newBlockReason.value = ''
-      
-      // Refresh activity log
-      await loadRecentActivity()
-    } catch (error) {
-      debug.error('Failed to block instance:', error)
-      alert('Failed to block instance. Check console for details.')
-    }
-  }
-}
-
-// eslint-disable-next-line unused-imports/no-unused-vars
-const unblockInstance = async (domain: string) => {
-  try {
-    await adminService.moderateInstance(
-      domain,
-      'unblock',
-      'Admin unblock',
-      authStore.session?.user?.id || ''
-    )
-
-    const index = blockedInstances.value.findIndex(i => i.domain === domain)
-    if (index !== -1) {
-      blockedInstances.value.splice(index, 1)
-    }
-    
-    // Refresh activity log
-    await loadRecentActivity()
-  } catch (error) {
-    debug.error('Failed to unblock instance:', error)
-    alert('Failed to unblock instance. Check console for details.')
-  }
-}
-
 // Utility functions
 const formatUptime = (timestamp: number) => {
   const diff = Date.now() - timestamp
@@ -858,102 +356,6 @@ const formatNumber = (num: number | undefined) => {
   return num.toString()
 }
 
-const formatTime = (date: Date) => {
-  return date.toLocaleTimeString()
-}
-
-const getActivityCategory = (type: string): string => {
-  if (!type) return 'other'
-  const t = type.toLowerCase()
-  if (t.startsWith('instance_') || t.includes('federation') || t.includes('add_instance')) return 'federation'
-  if (t.startsWith('user_') || t.includes('moderate') || t.includes('report') || t.includes('suspend')) return 'moderation'
-  if (t.includes('security') || t.includes('login') || t.includes('config')) return 'security'
-  return 'other'
-}
-
-const CONFIG_KEY_LABELS: Record<string, string> = {
-  instance_name: 'Instance name',
-  instance_description: 'Instance description',
-  terms_url: 'Terms URL',
-  privacy_url: 'Privacy URL',
-  max_post_length: 'Max post length',
-  max_server_size: 'Max server size',
-  max_message_length: 'Max message length',
-  max_custom_emojis_per_server: 'Max custom emojis per server',
-  custom_emoji_transform_quality: 'Custom emoji image quality',
-  allow_file_uploads: 'Allow file uploads',
-  enable_voice_channels: 'Enable voice channels',
-  federation_retry_attempts: 'Federation retry attempts',
-  oauth_providers: 'OAuth providers'
-}
-
-const formatConfigValue = (v: unknown): string => {
-  if (v == null || (typeof v === 'string' && v === '')) return '(empty)'
-  if (Array.isArray(v)) return v.join(', ')
-  if (typeof v === 'object') return JSON.stringify(v)
-  return String(v)
-}
-
-const formatActivityMessage = (event: { type: string; message: string | object; targetUsername?: string }) => {
-  const raw = event.message
-  const targetUser = event.targetUsername
-  if (raw == null || raw === '') {
-    if (event.type?.startsWith('user_') && targetUser) {
-      const verb = event.type === 'user_suspend' ? 'Suspended' : event.type === 'user_delete' ? 'Deleted' : event.type === 'user_unsuspend' ? 'Unsuspended' : event.type
-      return `${verb} @${targetUser}`
-    }
-    return event.type || 'Event'
-  }
-  try {
-    const obj = typeof raw === 'string' ? JSON.parse(raw) : raw
-    if (typeof obj !== 'object' || obj === null) return String(raw)
-
-    // Config change: key, new_value, old_value
-    if (obj.key != null && ('new_value' in obj || 'old_value' in obj)) {
-      const label = CONFIG_KEY_LABELS[obj.key] || obj.key.replace(/_/g, ' ')
-      const newVal = formatConfigValue(obj.new_value)
-      const oldVal = formatConfigValue(obj.old_value)
-      if (oldVal !== '(empty)' && newVal !== oldVal) {
-        return `${label}: ${newVal} (was ${oldVal})`
-      }
-      return `${label}: ${newVal}`
-    }
-
-    // User moderation with target
-    if ((event.type?.startsWith('user_') || obj.action === 'suspend' || obj.action === 'delete' || obj.action === 'unsuspend') && (targetUser || obj.user_id)) {
-      const who = targetUser ? `@${targetUser}` : (obj.user_id ? `user ${String(obj.user_id).slice(0, 8)}...` : '')
-      const reason = obj.reason ? ` - ${obj.reason}` : ''
-      const verb = event.type === 'user_suspend' ? 'Suspended' : event.type === 'user_delete' ? 'Deleted' : event.type === 'user_unsuspend' ? 'Unsuspended' : ''
-      return `${verb || 'Moderated'} ${who}${reason}`.trim()
-    }
-
-    // Build human-readable message from common keys
-    const parts: string[] = []
-    if (obj.domain) parts.push(obj.domain)
-    if (obj.reason) parts.push(`- ${obj.reason}`)
-    if (obj.action) parts.push(`(${obj.action})`)
-    if (obj.user_id && !targetUser) parts.push(`user: ${obj.user_id}`)
-    if (targetUser) parts.unshift(`@${targetUser}`)
-    if (parts.length) return parts.join(' ')
-
-    // Fallback: format key-value pairs
-    return Object.entries(obj)
-      .map(([k, v]) => `${k}: ${typeof v === 'object' ? JSON.stringify(v) : v}`)
-      .join(', ')
-  } catch {
-    return typeof raw === 'string' ? raw : String(raw)
-  }
-}
-
-const getActivityIcon = (type: string) => {
-  const cat = getActivityCategory(type)
-  switch (cat) {
-    case 'federation': return 'federation'
-    case 'security': return 'shield'
-    case 'moderation': return 'gavel'
-    default: return 'info'
-  }
-}
 
 </script>
 
@@ -966,6 +368,8 @@ const getActivityIcon = (type: string) => {
   font-family: var(--font-family);
   overflow-y: auto;
 }
+
+
 
 
 
@@ -988,11 +392,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .admin-title {
   display: flex;
   align-items: center;
   gap: 12px;
 }
+
+
 
 
 
@@ -1009,6 +417,8 @@ const getActivityIcon = (type: string) => {
   -webkit-text-fill-color: transparent;
   background-clip: text;
 }
+
+
 
 
 
@@ -1034,11 +444,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .system-status.healthy {
   background: rgba(0, 255, 136, 0.1);
   color: #00ff88;
   border: 1px solid rgba(0, 255, 136, 0.3);
 }
+
+
 
 
 
@@ -1058,11 +472,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .system-status.error {
   background: rgba(255, 69, 58, 0.1);
   color: #ff453a;
   border: 1px solid rgba(255, 69, 58, 0.3);
 }
+
+
 
 
 
@@ -1084,6 +502,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 @keyframes pulse {
   0%, 100% { opacity: 1; }
   50% { opacity: 0.5; }
@@ -1095,10 +515,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .admin-actions {
   display: flex;
   gap: 12px;
 }
+
+
 
 
 
@@ -1127,11 +551,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .action-btn:hover {
   background: var(--background-tertiary);
   border-color: var(--accent-color);
   transform: translateY(-1px);
 }
+
+
 
 
 
@@ -1151,12 +579,16 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .admin-grid {
   display: grid;
   /* min(600px, 100%) lets tracks collapse below 600px instead of overflowing */
   grid-template-columns: repeat(auto-fit, minmax(min(600px, 100%), 1fr));
   gap: 24px;
 }
+
+
 
 
 
@@ -1179,11 +611,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .admin-module:hover {
   border-color: var(--accent-color);
   box-shadow: 0 8px 32px rgba(0, 212, 255, 0.1);
   transform: translateY(-2px);
 }
+
+
 
 
 
@@ -1206,6 +642,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .module-header h2 {
   font-size: 18px;
   font-weight: 600;
@@ -1219,12 +657,16 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .stats-grid {
   display: grid;
   grid-template-columns: repeat(2, 1fr);
   gap: 16px;
   padding: 24px;
 }
+
+
 
 
 
@@ -1249,10 +691,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .stat-card:hover {
   border-color: var(--accent-color);
   box-shadow: 0 4px 16px rgba(0, 212, 255, 0.1);
 }
+
+
 
 
 
@@ -1273,9 +719,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .stat-content {
   flex: 1;
 }
+
+
 
 
 
@@ -1296,11 +746,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .stat-label {
   font-size: 14px;
   color: var(--text-secondary);
   margin-bottom: 4px;
 }
+
+
 
 
 
@@ -1319,9 +773,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .stat-change.positive {
   color: #00ff88;
 }
+
+
 
 
 
@@ -1333,6 +791,8 @@ const getActivityIcon = (type: string) => {
 .federation-content {
   padding: 24px;
 }
+
+
 
 
 
@@ -1357,6 +817,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .cyber-input:focus, .cyber-textarea:focus, .cyber-select:focus {
   outline: none;
   border-color: var(--accent-color);
@@ -1369,10 +831,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .cyber-textarea {
   resize: vertical;
   min-height: 80px;
 }
+
+
 
 
 
@@ -1396,6 +862,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 /* Override parent label styles so toggles stay horizontal and text doesn't truncate */
 .setting-group .toggle-label,
 .announcement-form .form-row.checks .toggle-label {
@@ -1409,9 +877,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .toggle-label .toggle-slider {
   flex-shrink: 0;
 }
+
+
 
 
 
@@ -1430,9 +902,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .toggle-label input[type="checkbox"] {
   display: none;
 }
+
+
 
 
 
@@ -1449,6 +925,8 @@ const getActivityIcon = (type: string) => {
   border-radius: 24px;
   transition: all 0.2s ease;
 }
+
+
 
 
 
@@ -1474,10 +952,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .toggle-label input[type="checkbox"]:checked + .toggle-slider {
   background: var(--accent-color);
   border-color: var(--accent-color);
 }
+
+
 
 
 
@@ -1496,10 +978,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 /* Blocked Instances */
 .blocked-instances {
   space-y: 16px;
 }
+
+
 
 
 
@@ -1524,10 +1010,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .instance-info .domain {
   font-weight: 600;
   color: var(--text-primary);
 }
+
+
 
 
 
@@ -1541,6 +1031,8 @@ const getActivityIcon = (type: string) => {
   color: var(--text-secondary);
   margin-top: 4px;
 }
+
+
 
 
 
@@ -1564,12 +1056,16 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 /* Federation Management Styles */
 .module-actions {
   display: flex;
   gap: 8px;
   margin-left: auto;
 }
+
+
 
 
 
@@ -1597,6 +1093,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .primary-btn-sm {
   padding: 6px 12px;
   font-size: 12px;
@@ -1608,10 +1106,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .primary-btn:hover, .primary-btn-sm:hover {
   background: #0099cc;
   transform: translateY(-1px);
 }
+
+
 
 
 
@@ -1635,6 +1137,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .meta-tag.domain {
   color: var(--harmony-primary, #7c8aff);
   background: rgba(124, 138, 255, 0.12);
@@ -1646,10 +1150,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .meta-tag.failures {
   color: #ff453a;
   background: rgba(255, 69, 58, 0.12);
 }
+
+
 
 
 
@@ -1673,11 +1181,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .status-indicator.ok {
   background: rgba(0, 255, 136, 0.1);
   border: 1px solid rgba(0, 255, 136, 0.3);
   color: #00ff88;
 }
+
+
 
 
 
@@ -1697,6 +1209,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .maintenance-card .action-btn.primary {
   background: var(--accent-primary);
   border-color: var(--accent-primary);
@@ -1708,9 +1222,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .spin {
   animation: spin 1s linear infinite;
 }
+
+
 
 
 
@@ -1729,9 +1247,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .error-text {
   color: #ff453a;
 }
+
+
 
 
 
@@ -1752,11 +1274,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .filter-controls {
   display: flex;
   gap: 12px;
   align-items: center;
 }
+
+
 
 
 
@@ -1779,6 +1305,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge {
   padding: 2px 8px;
   border-radius: 12px;
@@ -1786,6 +1314,8 @@ const getActivityIcon = (type: string) => {
   font-weight: 600;
   text-transform: uppercase;
 }
+
+
 
 
 
@@ -1804,10 +1334,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge.blocked {
   background: rgba(255, 69, 58, 0.2);
   color: #ff453a;
 }
+
+
 
 
 
@@ -1826,10 +1360,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge.success {
   background: rgba(0, 255, 136, 0.2);
   color: #00ff88;
 }
+
+
 
 
 
@@ -1848,9 +1386,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .detail-item {
   white-space: nowrap;
 }
+
+
 
 
 
@@ -1874,10 +1416,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .action-btn-sm:hover {
   border-color: var(--accent-color);
   color: var(--accent-color);
 }
+
+
 
 
 
@@ -1896,10 +1442,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .danger-btn-sm:hover {
   border-color: rgba(255, 69, 58, 0.5);
   color: #ff453a;
 }
+
+
 
 
 
@@ -1915,6 +1465,8 @@ const getActivityIcon = (type: string) => {
   padding: 20px;
   border-top: 1px solid var(--border-color);
 }
+
+
 
 
 
@@ -1938,10 +1490,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .pagination-btn:hover:not(:disabled) {
   border-color: var(--accent-color);
   color: var(--accent-color);
 }
+
+
 
 
 
@@ -1960,10 +1516,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .pagination-info {
   color: var(--text-secondary);
   font-size: 14px;
 }
+
+
 
 
 
@@ -1989,10 +1549,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .tab-btn.active {
   background: var(--accent-color);
   color: var(--text-primary);
 }
+
+
 
 
 
@@ -2017,6 +1581,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .detail-row {
   display: flex;
   gap: 8px;
@@ -2030,10 +1596,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .detail-row strong {
   min-width: 80px;
   color: var(--text-secondary);
 }
+
+
 
 
 
@@ -2056,9 +1626,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .spinning {
   animation: spin 1s linear infinite;
 }
+
+
 
 
 
@@ -2077,11 +1651,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .search-bar .icon {
   position: absolute;
   left: 12px;
   color: var(--text-secondary);
 }
+
+
 
 
 
@@ -2103,10 +1681,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge.suspended {
   background: rgba(255, 193, 7, 0.2);
   color: #ffc107;
 }
+
+
 
 
 
@@ -2125,10 +1707,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge.moderator {
   background: rgba(46, 204, 113, 0.2);
   color: #2ecc71;
 }
+
+
 
 
 
@@ -2150,6 +1736,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .metric-card {
   padding: 20px;
   background: var(--background-tertiary);
@@ -2163,9 +1751,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .metric-card.placeholder-metric {
   opacity: 0.5;
 }
+
+
 
 
 
@@ -2189,11 +1781,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .metric-status {
   width: 12px;
   height: 12px;
   border-radius: 50%;
 }
+
+
 
 
 
@@ -2211,6 +1807,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .metric-status.warning {
   background: #ffc107;
 }
@@ -2221,9 +1819,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .metric-status.error {
   background: #ff453a;
 }
+
+
 
 
 
@@ -2244,106 +1846,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .metric-detail {
   font-size: 12px;
   color: var(--text-secondary);
 }
 
 
-
-
-
-
-
-/* Activity Module */
-.activity-feed {
-  padding: 24px;
-  max-height: 400px;
-  overflow-y: auto;
-}
-
-
-
-
-
-
-
-.activity-item {
-  display: flex;
-  gap: 16px;
-  padding: 16px 0;
-  border-bottom: 1px solid var(--border-color);
-}
-
-
-
-
-
-
-
-.activity-item:last-child {
-  border-bottom: none;
-}
-
-
-
-
-
-
-
-.activity-icon {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 32px;
-  height: 32px;
-  border-radius: 50%;
-  font-size: 14px;
-}
-
-
-
-
-
-
-
-.activity-icon.federation {
-  background: rgba(0, 212, 255, 0.1);
-  color: #00d4ff;
-}
-
-
-
-
-
-
-
-.activity-icon.security {
-  background: rgba(255, 193, 7, 0.1);
-  color: #ffc107;
-}
-
-
-
-
-
-
-
-.activity-icon.moderation {
-  background: rgba(255, 69, 58, 0.1);
-  color: #ff453a;
-}
-
-
-
-
-
-
-
-.activity-icon.other {
-  background: rgba(128, 128, 128, 0.1);
-  color: var(--text-secondary);
-}
 
 
 
@@ -2366,27 +1876,11 @@ const getActivityIcon = (type: string) => {
 
 
 
-.announcement-form {
-  margin: 0 24px 20px;
-  padding: 20px;
-  background: var(--background-tertiary);
-  border-radius: 8px;
-  border: 1px solid var(--border-color);
-}
-
-
-
-
-
-
-.announcement-form h4 { margin: 0 0 16px 0; }
-
-
-
-
 
 
 .announcement-form .form-row { margin-bottom: 12px; }
+
+
 
 
 
@@ -2400,7 +1894,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .announcement-form .form-row.checks { display: flex; gap: 16px; flex-wrap: wrap; }
+
+
 
 
 
@@ -2419,8 +1917,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .announcement-form .form-row.two-col { grid-template-columns: 1fr; }
 }
+
+
 
 
 
@@ -2434,7 +1935,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .announcement-form .form-row.two-col label { margin-bottom: 4px; }
+
+
 
 
 
@@ -2453,51 +1958,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .announcement-form .form-actions { display: flex; gap: 8px; margin-top: 16px; }
 
 
-
-
-
-
-.announcements-list { padding: 0 24px 24px; }
-
-
-
-
-
-
-.announcement-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 12px 16px;
-  background: var(--background-tertiary);
-  border-radius: 8px;
-  margin-bottom: 8px;
-  border: 1px solid var(--border-color);
-}
-
-
-
-
-
-
-.announcement-meta { display: flex; align-items: center; gap: 8px; flex-wrap: wrap; }
-
-
-
-
-
-
-.announcement-icon { font-size: 18px; }
-
-
-
-
-
-
-.announcement-title { font-weight: 600; color: var(--text-primary); }
 
 
 
@@ -2507,11 +1972,6 @@ const getActivityIcon = (type: string) => {
 .announcement-item .badge.inactive { background: var(--background-quaternary); color: var(--text-muted); }
 
 
-
-
-
-
-.announcement-actions { display: flex; gap: 4px; }
 
 
 
@@ -2527,7 +1987,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .featured-servers-list { display: flex; flex-direction: column; gap: 8px; padding: 0 24px 24px; }
+
+
 
 
 
@@ -2550,7 +2014,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .featured-server-item:hover { border-color: var(--accent-color); }
+
+
 
 
 
@@ -2561,6 +2029,8 @@ const getActivityIcon = (type: string) => {
   border-color: rgba(255, 193, 7, 0.5);
   background: rgba(255, 193, 7, 0.05);
 }
+
+
 
 
 
@@ -2579,12 +2049,16 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .server-icon {
   width: 100%;
   height: 100%;
   border-radius: 8px;
   object-fit: cover;
 }
+
+
 
 
 
@@ -2606,7 +2080,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .server-details { flex: 1; min-width: 0; }
+
+
 
 
 
@@ -2620,14 +2098,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .server-details .server-meta { font-size: 13px; color: var(--text-secondary); }
 
 
-
-
-
-
-.featured-order { margin-left: 8px; opacity: 0.8; }
 
 
 
@@ -2641,7 +2116,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .featured-server-item .action-btn-sm.unpin-btn { color: var(--text-secondary); }
+
+
 
 
 
@@ -2651,39 +2130,6 @@ const getActivityIcon = (type: string) => {
 .featured-server-item .action-btn-sm { display: flex; align-items: center; gap: 6px; }
 
 
-
-
-
-
-
-.activity-content {
-  flex: 1;
-}
-
-
-
-
-
-
-
-.activity-message {
-  font-size: 14px;
-  color: var(--text-primary);
-  margin-bottom: 4px;
-}
-
-
-
-
-
-
-
-.activity-meta {
-  display: flex;
-  gap: 12px;
-  font-size: 12px;
-  color: var(--text-secondary);
-}
 
 
 
@@ -2703,10 +2149,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .emoji-content {
   padding: 0;
   overflow-y: auto;
 }
+
+
 
 
 
@@ -2725,11 +2175,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .performance-content {
   padding: 0;
   max-height: 800px;
   overflow-y: auto;
 }
+
+
 
 
 
@@ -2759,6 +2213,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .config-tab-btn:hover {
   color: var(--text-primary);
   background: var(--background-tertiary);
@@ -2771,10 +2227,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .config-tab-btn.active {
   color: var(--accent-color);
   border-bottom-color: var(--accent-color);
 }
+
+
 
 
 
@@ -2801,6 +2261,7 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .admin-grid {
     display: flex;
     flex-direction: column;
@@ -2810,15 +2271,18 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .admin-module {
     max-width: calc(100vw - 32px);
   }
 
 
 
+
   .admin-panel {
     padding: 16px;
   }
+
 
 
 
@@ -2832,6 +2296,7 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .admin-title {
     flex-direction: column;
     align-items: flex-start;
@@ -2841,10 +2306,12 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   /* Two-up stats read better than a single tall column on phones. */
   .stats-grid {
     grid-template-columns: 1fr 1fr;
   }
+
 
 
 
@@ -2858,9 +2325,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .admin-panel {
     padding: 12px;
   }
+
 
 
 
@@ -2870,9 +2339,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .stats-grid {
     grid-template-columns: 1fr;
   }
+
 
 
 
@@ -2884,11 +2355,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
   .admin-actions .action-btn {
     width: 100%;
     justify-content: center;
     min-height: 44px;
   }
+
 
 
 
@@ -2902,6 +2375,8 @@ const getActivityIcon = (type: string) => {
     -webkit-overflow-scrolling: touch;
   }
 }
+
+
 
 
 
@@ -2923,11 +2398,15 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .server-icon img {
   width: 100%;
   height: 100%;
   object-fit: cover;
 }
+
+
 
 
 
@@ -2943,6 +2422,8 @@ const getActivityIcon = (type: string) => {
   align-items: center;
   gap: 8px;
 }
+
+
 
 
 
@@ -2966,12 +2447,16 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .server-meta {
   display: flex;
   gap: 16px;
   font-size: 12px;
   color: var(--text-secondary);
 }
+
+
 
 
 
@@ -2986,7 +2471,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .report-type-badge.post { background: rgba(87, 242, 135, 0.2); color: #57f287; }
+
+
 
 
 
@@ -3000,7 +2489,11 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .report-type-badge.server { background: rgba(235, 69, 158, 0.2); color: #eb459e; }
+
+
 
 
 
@@ -3020,9 +2513,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .report-proof :deep(.report-link:hover) {
   opacity: 0.8;
 }
+
+
 
 
 
@@ -3041,10 +2538,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge.sensitive {
   background: rgba(250, 166, 26, 0.2);
   color: #faa61a;
 }
+
+
 
 
 
@@ -3063,10 +2564,14 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .badge.silenced {
   background: rgba(250, 166, 26, 0.2);
   color: #faa61a;
 }
+
+
 
 
 
@@ -3085,9 +2590,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .mod-btn.warning-btn:hover {
   background: rgba(250, 166, 26, 0.3);
 }
+
+
 
 
 
@@ -3108,6 +2617,8 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .add-tier-form .cyber-input {
   flex: 1;
   min-width: 100px;
@@ -3119,9 +2630,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .funding-link-row .cyber-input {
   min-width: 0;
 }
+
+
 
 
 
@@ -3140,9 +2655,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .supporter-search-wrapper .cyber-input {
   width: 100%;
 }
+
+
 
 
 
@@ -3161,9 +2680,13 @@ const getActivityIcon = (type: string) => {
 
 
 
+
+
 .supporter-suggestion-item.selected .supporter-suggestion-handle {
   color: rgba(255, 255, 255, 0.6);
 }
+
+
